ENGR00214328 make jpegProcess support more format than HAL_PIXEL_FORMAT_BLOB.
jpegProcess only support HAL_PIXEL_FORMAT_BLOB format when take picture.
Because the format HAL_PIXEL_FORMAT_BLOB is a new customized format in android 4.2.
And now, the gralloc does not support this format.
So, extend it to support more format.

In JpegProcessor::processNewCapture, captureBuffer is defined two times.
One in { }, another in whole function.
But captureBuffer set value in {}, but used outside {}.

This commit merge with fef8536 ENGR00318735 can't take picture on kk4.4.3_3.10.y.

Conflicts:
	services/camera/libcameraservice/api1/client2/JpegProcessor.cpp

Change-Id: I54f63a588d0ed1a5d8eaacbf50df46036b592067
Signed-off-by: Xiaowen Liu <b37945@freescale.com>
Signed-off-by: Fang Hui <b31070@freescale.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/597
Reviewed-by: fang hui <hui.fang@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/995